adjust the structure of the files and add a banner image for the readme file
