--- burguer-queen-react/src/babel.config.js
@@ -1,8 +1,0 @@
-export default {
-    presets: [
-      [
-        '@babel/preset-env',
-        '@babel/preset-react'
-      ],
-    ],
-  };
