--- burguer-queen-react/src/jest.config.js
@@ -1,6 +1,0 @@
-export default {
-  testEnvironment: 'jsdom',
-  transform: {
-      '^.+\\.jsx?$': 'babel-jest', // transpile mjs, mts, js, ts files
-  },
-};
