+++ src/babel.config.js
@@ -1,0 +1,8 @@
+export default {
+    presets: [
+      [
+        '@babel/preset-env',
+        '@babel/preset-react'
+      ],
+    ],
+  };
