+++ src/componentes/waiter/Waiter.css
@@ -1,0 +1,347 @@
+/* Estilos para el logo */
+@font-face {
+  font-family: 'Hey Comic';
+  src: url('../fonts/Hey Comic.ttf') format('opentype');
+}
+.waiter-logoburger-img {
+    width: 200px;
+    height: 200px;
+  }
+ 
+  .login-background {
+    /* background: linear-gradient(to bottom, #d7ffff 0%, #bdf0f7 25%, #9edde8 50%, #7ecad9 75%, #61b9cd 100%); */
+   background-color:rgb(5, 5, 5) ;
+  }
+  
+  .others-background{
+    background-color: #F8F0D7 ;
+  }
+  /* Estilos para la barra de navegación */
+  nav ul {
+    list-style: none;
+    display: flex;
+    gap: 20px;
+  }
+  
+  nav ul li {
+    display: inline;
+  }
+  
+  nav ul li a {
+    text-decoration: none;
+    color: #724C11;
+    font-weight: bold;
+  }
+  
+  
+::-webkit-scrollbar {
+  width: 5px;
+}
+
+::-webkit-scrollbar-track {
+  background-color: #f5f5f5;
+}
+
+::-webkit-scrollbar-thumb {
+  background-color: #724C11;
+}
+  
+  /* Estilos para el botón de logout */
+  .waiter-btn {
+    background-color: #724C11;
+    color: #fff;
+    padding: 10px 20px;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
+  
+  .waiter-btn:hover {
+    background-color: #555;
+  }
+
+  .wt-main-container{
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    margin-top: 50px;
+  }
+
+  .wt-products-section{
+    display: flex;
+    flex-direction: column;
+    width: 50%;
+    align-items: center;
+    /* height: 500px;
+    overflow: auto; */
+  }
+
+  .wt-products_section__buttons{
+ display:  flex;
+ justify-content: center;
+  }
+  .wt-products-section__button{
+    margin-right: 10px;
+    font-family: 'Roboto';
+    font-style: normal;
+    font-size: 15px;
+    text-align: center;
+    color: #FFFFFF;
+    width: 159.69px;
+    height: 46px;
+    background: #724C11;
+    border-radius: 16px;
+  }
+  .wt-products-section__content{
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: space-evenly;
+    align-items: center;
+    font-family: "Rubik";
+    margin-top: 4%;
+    margin-bottom: 5%;
+    padding: 30px 40px;
+    height: 450px;
+    overflow: auto;
+  }
+  /* .wt-card{
+ 
+  } */
+  .card-img {
+    justify-content: center;
+    background: #fff9eb;
+    width: 85px;
+    height: 84px;
+    object-fit: cover;
+    margin-top: 10%;
+    display: flex;
+    align-items: center;
+}
+.card-img img{
+  width: 50px;
+    height: 50px;
+}
+
+  .card-of-product{
+    background-color: #fff;
+  box-shadow: 0 .5rem 1rem rgba(0,0,0,.15)!important;
+  align-items: center;
+  flex-direction: column;
+  display: flex;
+  width: 30%;
+  height: 60%;
+  justify-content: center;
+  border-radius: 0.75rem;
+  margin-bottom: 5%;
+  }
+  .wt-order-section {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    width: 50%;
+  }
+  .wt-orders-table__header-container{
+    margin-bottom: 5%;
+    width: 100%;
+    display: flex;
+    flex-direction: column;
+  }
+  .wt-orders-table__table {
+    width: 100%;
+    border-collapse: collapse;
+  }
+  
+  .wt-orders-table__heading {
+    background-color: #f2f2f2;
+    padding: 10px;
+    text-align: left;
+  }
+  
+  .wt-orders-table__cell {
+    padding: 10px;
+    border-bottom: 1px solid #ccc;
+    border: 1px solid black;
+    text-align: left;
+  }
+  
+  .wt-orders-table__image {
+    width: 50px;
+    height: 50px;
+  }
+  
+  .wt-orders-table__price {
+    margin: 0;
+  }
+  
+  .wt-orders-table__button {
+    padding: 5px 10px;
+    margin-right: 5px;
+  }
+
+  .buttonDecrement:active{
+    background-color: #efe9458a;
+  }
+
+  .buttonIncrease:active{
+    background-color: rgba(0, 128, 0, 0.42);
+  }
+
+  .buttonCancel:active{
+    background-color: rgba(232, 43, 43, 0.557);
+  }
+  
+  .wt-orders-table__row:nth-child(even) {
+    background-color: #f9f9f9;
+  }
+/* .wt-orders-table__total{
+
+} */
+.wt-orders-table{
+  background-color: rgb(255, 255, 255);
+  padding: 30px 40px;
+  width: 50%;
+  border-radius: 20px;
+}
+/* .wt-orders-table__submit-button{
+
+} */
+
+.wt-orders-table__label {
+  display: inline-block;
+  vertical-align: middle;
+  margin-right: 10px;
+}
+
+.wt-orders-table__input {
+  border:  none;
+  display: inline-block;
+  vertical-align: middle;
+    height: 26px;
+    background: #F4F3E4;
+    border-radius: 5px;
+}
+
+.wt-orders-table__table-container {
+  max-height: 400px; /* ajusta la altura máxima según tus necesidades */
+  overflow-y: auto;
+  width: 100%;/* establece el ancho al 100% */
+
+}
+
+.wt-orders-table__table {
+  width: 100%;
+  table-layout: fixed;
+  white-space: nowrap;
+}
+
+.wt-orders-table__thead-container {
+  position: sticky;
+  top: 0;
+  background-color: #fff;
+  z-index: 1;
+  display: table;
+  width:50%;
+}
+
+.wt-orders-table__thead {
+  display: table;
+  width: 100%;
+  table-layout: fixed;
+}
+
+.wt-orders-table__tbody {
+  width: 80%;
+    position: sticky;
+
+  display: table;
+  max-height: 500px; /* ajusta la altura máxima según tus necesidades */
+  overflow-y: auto;
+}
+
+.wt-orders-table__heading {
+  white-space: nowrap;
+}
+
+.wt-orders-table__total {
+  margin-top: 10px;
+}
+
+.wt-orders-table__submit-button {
+  margin-right: 10px;
+  font-family: 'Roboto';
+  font-style: normal;
+  font-size: 15px;
+  text-align: center;
+  color: #FFFFFF;
+  width: 159.69px;
+  height: 46px;
+  background: #724C11;
+  border-radius: 16px;
+  margin-top: 10px;
+}
+.card {
+  box-shadow: 0 .5rem 1rem rgba(0,0,0,.15)!important;
+  background: white;
+  border: 1px solid #ccc;
+  border-radius: 20px;
+  padding: 40px;
+  margin-bottom: 20px;
+  height: 282px;
+  width: 396px;
+  overflow: auto;
+}
+.card h2 {
+  font-family: "Hey Comic";
+  margin-top: 0;
+  text-align: center;
+}
+.info {
+  margin-bottom: 10px;
+}
+table {
+  width: 100%;
+  border-collapse: collapse;
+}
+th, td {
+  padding: 8px;
+  border-bottom: 1px solid #ddd;
+}
+th {
+  text-align: left;
+}
+.total {
+  font-weight: bold;
+}
+.ordersList-main{
+    display: flex;
+flex-wrap: wrap;
+justify-content: space-evenly;
+align-items: center;
+font-family: "Rubik";
+margin-top: 50px;
+margin-bottom: 5%;
+}
+@media (min-width: 1024px) and (max-height: 768px) {
+  /* Estilos específicos para dispositivos en modo horizontal */
+  /* Aquí puedes agregar tus estilos */
+  .wt-products-section__content {
+    height: 350px;
+    overflow: auto;
+  }
+  .wt-orders-table__table-container {
+    max-height: 200px;
+    overflow-y: auto;
+    width: 100%;
+}
+.wt-orders-table {
+  background-color: rgb(255, 255, 255);
+  padding: 30px 40px;
+  width: 60%;
+  border-radius: 20px;
+}
+.ordersList-main{
+  height: 500px;
+  overflow: auto;
+}
+}
