+++ src/jest.config.js
@@ -1,0 +1,6 @@
+export default {
+  testEnvironment: 'jsdom',
+  transform: {
+      '^.+\\.jsx?$': 'babel-jest', // transpile mjs, mts, js, ts files
+  },
+};
